refactoring component tree - added noise menu
