+++ src/styles/components/_noise-menu.scss
@@ -1,0 +1,13 @@
+.noise-menu {
+  width: 100%;
+  height: 90%;
+}
+
+.noises {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
