+++ src/styles/components/_rootapp.scss
@@ -8,11 +8,3 @@
   height: 100%;
 }
 
-.noises {
-  width: 100%;
-  height: 90%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
