+++ src/styles/styles.scss
@@ -6,6 +6,7 @@
 @import './base/animations';
 //COMPONENTS---------------
 @import './components/rootapp';
+@import './components/noise-menu';
 @import './components/noise';
 @import './components/footer'
